Switch to discord from slack, add snooze
Slack was not properly notifying on my phone, also the slack clients
available seemed rather lackluster. After a little more research I found
serenity as an alternative to discord-rs (which depends on an old
version of the ssl crate which isn't compatible with openssl 1.1).

In addition to the transfer from slack to discord interfaces, I've added
a snooze feature to complement the nag feature. Now it should be
possible to tell the bot to !shutup, which will make it leave you alone
for a few hours.